Null許容型にした影響のワーニングの修正

+++ ExtremeRoles/Roles/API/Interface/IRoleSpecialReset.cs
@@ -35,7 +35,10 @@ public interface IRoleSpecialReset
 
     public static void ResetRole(byte targetPlayerId)
     {
-        PlayerControl resetPlayer = Player.GetPlayerControlById(targetPlayerId);
+		if (!Player.TryGetPlayerControl(targetPlayerId, out var resetPlayer))
+		{
+			return;
+		}
         SingleRoleBase resetRole = ExtremeRoleManager.GameRole[targetPlayerId];
 
         IParentChainStatus.PurgeParent(targetPlayerId);
